# Performance update for the exif package