--- a/exif/gps.go
+++ b/exif/gps.go
@@ -1,1 +1,0 @@
-package exif