… .github/workflows/go.yml
… exif/benchmark_test.go
… exif/exif.go
… exif/exif_test.go
… exif/exifdata.go
… exif/ifds/gpsifd/gpsIfd.go
… exif/ifds/ifds.go
… exif/ifds/ifds_test.go
… exif/ifds/mknote/mknote_test.go
… exif/ifds/mknote/nikon.go
… exif/parser.go
… exif/parser_test.go
… exif/reader.go
… exif/tag/parse.go
… exif/tag/parse_test.go
… exif/tag/tag.go
… exif/tag/tagType.go
… exif/tag/tag_test.go
… exif/thumbnail.go
… exif/tiffTagEnumerator.go
… exif/time.go
… exif/types.go
… go.mod
… go.sum
… heic/heic.go
… heic/heic_test.go
… imagemeta.go
… meta/exifTypes.go
… meta/exifTypes_test.go
… test/main.go