Refactored the code with the goal of making it easier to read and increased performance.

Benchmarks for exif parsing show ~ 35 - 45%  improvement.
```
name                 old time/op    new time/op    delta
ScanExif100/1.CR2      36.2µs ± 7%    21.6µs ± 5%  -40.33%  (p=0.000 n=19+19)
ScanExif100/3.CR2      31.3µs ± 6%    16.5µs ±19%  -47.38%  (p=0.000 n=19+20)
ScanExif100/60D.CR2    33.9µs ±13%    21.3µs ± 8%  -37.22%  (p=0.000 n=20+20)
ScanExif100/6D.CR2     35.8µs ± 9%    23.2µs ± 9%  -35.26%  (p=0.000 n=19+20)
ScanExif100/7D.CR2     36.4µs ± 9%    21.4µs ± 6%  -41.16%  (p=0.000 n=20+20)
ScanExif100/90D.cr3    6.30µs ± 8%    3.83µs ± 9%  -39.29%  (p=0.000 n=20+19)
ScanExif100/2.CR3      6.23µs ± 6%    4.01µs ±24%  -35.60%  (p=0.000 n=19+20)
ScanExif100/1.CR3      6.21µs ± 7%    4.17µs ±44%  -32.87%  (p=0.000 n=19+19)
ScanExif100/1.jpg      13.7µs ± 6%     8.0µs ±14%  -41.82%  (p=0.000 n=19+20)
ScanExif100/2.jpg      17.6µs ± 5%    12.3µs ±29%  -30.22%  (p=0.000 n=18+19)

name                 old alloc/op   new alloc/op   delta
ScanExif100/1.CR2      13.2kB ± 0%    11.1kB ± 0%  -15.56%  (p=0.000 n=19+20)
ScanExif100/3.CR2      12.1kB ± 0%     5.7kB ± 0%  -52.76%  (p=0.000 n=19+19)
ScanExif100/60D.CR2    12.6kB ± 0%    11.1kB ± 0%  -11.78%  (p=0.000 n=20+20)
ScanExif100/6D.CR2     13.2kB ± 0%    11.1kB ± 0%  -15.47%  (p=0.000 n=20+18)
ScanExif100/7D.CR2     13.2kB ± 0%    11.1kB ± 0%  -15.56%  (p=0.000 n=20+19)
ScanExif100/90D.cr3    5.13kB ± 0%    0.98kB ± 0%  -80.84%  (p=0.000 n=20+20)
ScanExif100/2.CR3      5.12kB ± 0%    0.98kB ± 0%  -80.96%  (p=0.000 n=16+20)
ScanExif100/1.CR3      5.13kB ± 0%    0.98kB ± 0%  -80.96%  (p=0.000 n=20+20)
ScanExif100/1.jpg      6.72kB ± 0%    2.58kB ± 0%  -61.64%  (p=0.000 n=20+17)
ScanExif100/2.jpg      8.39kB ± 0%    5.43kB ± 0%  -35.24%  (p=0.000 n=18+16)

name                 old allocs/op  new allocs/op  delta
ScanExif100/1.CR2        56.0 ± 0%      31.0 ± 0%  -44.64%  (p=0.000 n=20+20)
ScanExif100/3.CR2        52.0 ± 0%      29.0 ± 0%  -44.23%  (p=0.000 n=20+20)
ScanExif100/60D.CR2      51.0 ± 0%      29.0 ± 0%  -43.14%  (p=0.000 n=20+20)
ScanExif100/6D.CR2       56.0 ± 0%      31.0 ± 0%  -44.64%  (p=0.000 n=20+20)
ScanExif100/7D.CR2       56.0 ± 0%      31.0 ± 0%  -44.64%  (p=0.000 n=20+20)
ScanExif100/90D.cr3      16.0 ± 0%       9.0 ± 0%  -43.75%  (p=0.000 n=20+20)
ScanExif100/2.CR3        16.0 ± 0%       9.0 ± 0%  -43.75%  (p=0.000 n=20+20)
ScanExif100/1.CR3        16.0 ± 0%       9.0 ± 0%  -43.75%  (p=0.000 n=20+20)
ScanExif100/1.jpg        29.0 ± 0%      18.0 ± 0%  -37.93%  (p=0.000 n=20+20)
ScanExif100/2.jpg        32.0 ± 0%      23.0 ± 0%  -28.12%  (p=0.000 n=20+20)
```

Benchmarks for tag extraction have reduced allocations and B/op.
```
// GPSDate
// BenchmarkOld/2.CR2         	 1000000	      1044 ns/op	     176 B/op	       4 allocs/op
// BenchmarkNew/2.CR2         	 4983446	       272 ns/op	       0 B/op	       0 allocs/op

// GPSCoords
// BenchmarkOld/2.CR2         	 1340478	       850 ns/op	     224 B/op	       4 allocs/op
// BenchmarkNew/2.CR2         	 3894754	       303 ns/op	       0 B/op	       0 allocs/op

// Aperture
// BenchmarkOld/0.CR2         	 3556357	       303 ns/op	      80 B/op	       1 allocs/op
// BenchmarkNew/0.CR2         	10543807	       123 ns/op	       0 B/op	       0 allocs/op

// DateTime
// BenchmarkOld/0.CR2         	 1000000	      1043 ns/op	     208 B/op	       5 allocs/op
// BenchmarkNew/0.CR2         	 6169940	       230 ns/op	       0 B/op	       0 allocs/op
```